Tag parse tree with coref tags
Addresses #2
This commit uses a right to left BFS of the tree first to find coref
tags per sentence. If the coref tag cannot be found in the sentence
using BFS, it then tries a regex to insert the tag. Has been shown to
find about 93% of corefs and properly tag them in the tree.